requirements.txt
fuzzywuzzy==0.17.0
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
PyYAML==5.1
six==1.12.0
